Update DetailActivity.java
Create a shareIntent method in Details.
Created a share intent with Intent action Send
Set the intent type to text/plain
Created a sharestring that is a combination of various predefined strings that are already defined in the class.
We use the intent method putextra to put the string up for the activity that will pick it up.
Lastly we start our shareintent with startActivity method